add option to clear data
Now it is possible to clear all data using "Clear data" button in "Advanced" tab.
It should help to start over with a clean slate if the app doesn't work properly.
It also allows you to delete the default deck and model created after the first start. So you can clear data, set your own settings with UI and then restart the app to apply changes.